LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/sandbox-top-navigation-grandchild-unsandboxed-inherit.tentative.sub.window-expected.txt


FAIL An unsandboxed grandchild inherits its parents ability to navigate top. assert_equals: The navigation should succeed. expected true but got false
